adding updated notebook with CNN
Notebook with an incredibly simple CNN structure that I got from PyTorch's tutorial. Got a very biased result at the end, because I trained the network on a very small dataset. Next step is to train it on a bigger dataset and hopefully improve the results. Also want to implement a better architecture - will start with something resembling LeNet.